move to test_timeline_detach_ancestor
